Remove code duplication
Avoid code duplication with class inheritance.

Methods that are used in multiple test classes across both advocate and
litigator are in `base_class_test.rb` while those used in multiple tests
within either the advocate or litigator group are in the appropriate `base.rb`.
Methods that are only used in a single test class are kept in that file.
