+++ .rubocop_todo.yml
@@ -201,7 +201,6 @@ Metrics/ClassLength:
     - 'app/validators/claim/interim_claim_validator.rb'
     - 'app/validators/fee/base_fee_validator.rb'
     - 'lib/geckoboard_publisher/travel_automation_report.rb'
-    - 'spec/support/api/claims/base_claim_test.rb'
 
 # Offense count: 14
 # Configuration parameters: AllowedMethods, AllowedPatterns, IgnoredMethods, Max.
@@ -291,15 +290,15 @@ Metrics/MethodLength:
     - 'spec/presenters/interim_fee_type_presenter_spec.rb'
     - 'spec/services/stats/claim_creation_source_data_generator_spec.rb'
     - 'spec/services/stats/claim_submissions_data_generator_spec.rb'
-    - 'spec/support/api/claims/advocate_claim_test.rb'
-    - 'spec/support/api/claims/advocate_hardship_claim_test.rb'
-    - 'spec/support/api/claims/advocate_interim_claim_test.rb'
-    - 'spec/support/api/claims/advocate_supplementary_claim_test.rb'
+    - 'spec/support/api/claims/advocate_claim_test/final.rb'
+    - 'spec/support/api/claims/advocate_claim_test/hardship.rb'
+    - 'spec/support/api/claims/advocate_claim_test/interim.rb'
+    - 'spec/support/api/claims/advocate_claim_test/supplementary.rb'
     - 'spec/support/api/claims/base_claim_test.rb'
-    - 'spec/support/api/claims/litigator_final_claim_test.rb'
-    - 'spec/support/api/claims/litigator_hardship_claim.rb'
-    - 'spec/support/api/claims/litigator_interim_claim_test.rb'
-    - 'spec/support/api/claims/litigator_transfer_claim_test.rb'
+    - 'spec/support/api/claims/litigator_claim_test/final.rb'
+    - 'spec/support/api/claims/litigator_claim_test/hardship.rb'
+    - 'spec/support/api/claims/litigator_claim_test/interim.rb'
+    - 'spec/support/api/claims/litigator_claim_test/transfer.rb'
     - 'spec/support/factory_helpers.rb'
     - 'spec/support/matchers/hash_matchers.rb'
     - 'spec/support/scheme_date_helpers.rb'
