+++ spec/support/api/claims/advocate_claim_test/supplementary.rb
@@ -1,58 +1,50 @@
-require_relative 'base_claim_test'
+require 'advocate_claim_test/base'
 
-class AdvocateSupplementaryClaimTest < BaseClaimTest
-  def initialize(...)
-    @claim_create_endpoint = 'claims/advocates/supplementary'
-    @email = ADVOCATE_TEST_EMAIL
-    @role = 'agfs'
+module AdvocateClaimTest
+  class Supplementary < Base
+    def initialize(...)
+      @claim_create_endpoint = 'claims/advocates/supplementary'
 
-    super
-  end
+      super
+    end
 
-  def test_creation!
-    super
+    def test_creation!
+      super
 
-    # CREATE miscellaneous fee
-    response = @client.post_to_endpoint('fees', misc_fee_data)
+      # CREATE miscellaneous fee
+      response = @client.post_to_endpoint('fees', misc_fee_data)
 
-    # add date attended to miscellaneous fee
-    @attended_item_id = response['id']
-    @client.post_to_endpoint('dates_attended', date_attended_data)
+      # add date attended to miscellaneous fee
+      @attended_item_id = response['id']
+      @client.post_to_endpoint('dates_attended', date_attended_data)
 
-    # add expense
-    @client.post_to_endpoint('expenses', expense_data)
-  ensure
-    clean_up
-  end
+      # add expense
+      @client.post_to_endpoint('expenses', expense_data)
+    ensure
+      clean_up
+    end
 
-  def claim_data
-    advocate_category = fetch_value(ADVOCATE_CATEGORY_ENDPOINT)
-    court_id = fetch_id(COURT_ENDPOINT)
+    def claim_data
+      advocate_category = fetch_value(ADVOCATE_CATEGORY_ENDPOINT)
+      court_id = fetch_id(COURT_ENDPOINT)
 
-    {
-      api_key:,
-      creator_email: 'advocateadmin@example.com',
-      advocate_email: 'advocate@example.com',
-      case_number: 'B20161234',
-      providers_ref: SecureRandom.uuid[3..15].upcase,
-      advocate_category:,
-      court_id:,
-      cms_number: '12345678',
-      additional_information: 'string',
-      apply_vat: true
-    }
-  end
+      {
+        api_key:,
+        creator_email: 'advocateadmin@example.com',
+        advocate_email: 'advocate@example.com',
+        case_number: 'B20161234',
+        providers_ref: SecureRandom.uuid[3..15].upcase,
+        advocate_category:,
+        court_id:,
+        cms_number: '12345678',
+        additional_information: 'string',
+        apply_vat: true
+      }
+    end
 
-  def misc_fee_data
-    # Only certain misc fees are eligible e.g. Confiscation hearings (half day) - MIDTH
-    fee_type_id = fetch_id(FEE_TYPE_ENDPOINT, category: 'misc', role: 'agfs', unique_code: 'MIDTH')
-
-    {
-      api_key:,
-      claim_id: @claim_uuid,
-      fee_type_id:,
-      quantity: 2,
-      rate: 1.55
-    }
+    def fee_type_options
+      # Only certain misc fees are eligible e.g. Confiscation hearings (half day) - MIDTH
+      super.merge(unique_code: 'MIDTH')
+    end
   end
 end
