+++ spec/support/api/claims/litigator_claim_test/interim.rb
@@ -1,55 +1,55 @@
-require_relative 'base_claim_test'
+require 'litigator_claim_test/base'
 
-class LitigatorInterimClaimTest < BaseClaimTest
-  def initialize(...)
-    @claim_create_endpoint = 'claims/interim'
-    @email = LITIGATOR_TEST_EMAIL
-    @role = 'lgfs'
+module LitigatorClaimTest
+  class Interim < Base
+    def initialize(...)
+      @claim_create_endpoint = 'claims/interim'
 
-    super
-  end
+      super
+    end
 
-  def test_creation!
-    super
+    def test_creation!
+      super
 
-    # CREATE interim fee
-    @client.post_to_endpoint('fees', interim_fee_data)
+      # CREATE interim fee
+      @client.post_to_endpoint('fees', interim_fee_data)
 
-    # CREATE a disbursement
-    @client.post_to_endpoint('disbursements', disbursement_data)
-  ensure
-    clean_up
-  end
+      # CREATE a disbursement
+      @client.post_to_endpoint('disbursements', disbursement_data)
+    ensure
+      clean_up
+    end
 
-  def claim_data
-    case_type_id = fetch_id(CASE_TYPE_ENDPOINT, index: 12, role: 'lgfs') # Trial
-    offence_id = fetch_id(OFFENCE_ENDPOINT, offence_description: 'Miscellaneous/other')
-    court_id = fetch_id(COURT_ENDPOINT)
+    def claim_data
+      case_type_id = fetch_id(CASE_TYPE_ENDPOINT, index: 12, role: 'lgfs') # Trial
+      offence_id = fetch_id(OFFENCE_ENDPOINT, offence_description: 'Miscellaneous/other')
+      court_id = fetch_id(COURT_ENDPOINT)
 
-    {
-      api_key:,
-      creator_email: 'litigatoradmin@example.com',
-      user_email: 'litigator@example.com',
-      case_number: 'A20161234',
-      supplier_number:,
-      case_type_id:,
-      offence_id:,
-      court_id:,
-      cms_number: '12345678',
-      additional_information: 'string',
-      effective_pcmh_date: 1.month.ago.as_json
-    }
-  end
+      {
+        api_key:,
+        creator_email: 'litigatoradmin@example.com',
+        user_email: 'litigator@example.com',
+        case_number: 'A20161234',
+        supplier_number:,
+        case_type_id:,
+        offence_id:,
+        court_id:,
+        cms_number: '12345678',
+        additional_information: 'string',
+        effective_pcmh_date: 1.month.ago.as_json
+      }
+    end
 
-  def interim_fee_data
-    fee_type_id = fetch_id(FEE_TYPE_ENDPOINT, index: 1, category: 'interim') # Effective PCMH
+    def interim_fee_data
+      fee_type_id = fetch_id(FEE_TYPE_ENDPOINT, index: 1, category: 'interim') # Effective PCMH
 
-    {
-      api_key:,
-      claim_id: @claim_uuid,
-      fee_type_id:,
-      quantity: 5,
-      amount: 200.50
-    }
+      {
+        api_key:,
+        claim_id: @claim_uuid,
+        fee_type_id:,
+        quantity: 5,
+        amount: 200.50
+      }
+    end
   end
 end
